Storm Surge safty, moving town fn to extras
