+++ commands/Game/town.js
@@ -2,15 +2,58 @@ const { SlashCommandBuilder, EmbedBuilder, ActionRowBuilder, ButtonBuilder, Butt
 
 const { Town, MediumTile, GuildData, UserData, MaterialStore, TownMaterial, TownPlots, PlayerBuilding, CoreBuilding } = require('../../dbObjects.js');
 
-const { loadBuilding } = require('./exported/displayBuilding.js');
+// const { loadBuilding } = require('./exported/displayBuilding.js');
 
-const coreReq = require('../../events/Models/json_prefabs/coreBuildings.json');
-const coreBuildCostList = require('../Development/Export/Json/coreBuildCostList.json');
-const { checkUserTownPerms, checkUserAsMayor, grabUser, grabTown, grabTownByName, makeCapital, createInteractiveChannelMessage, editTimedChannelMessage, sendTimedChannelMessage, grabLocalTowns, handleLimitOnOptions, createConfirmCancelButtonRow, randArrPos, handleCatchDelete } = require('../../uniHelperFunctions.js');
+// const coreReq = require('../../events/Models/json_prefabs/coreBuildings.json');
+// const coreBuildCostList = require('../Development/Export/Json/coreBuildCostList.json');
+const { 
+	checkUserTownPerms, 
+	checkUserAsMayor, 
+	grabUser, 
+	grabTown, 
+	grabTownByName, 
+	makeCapital, 
+	createInteractiveChannelMessage, 
+	editTimedChannelMessage, 
+	sendTimedChannelMessage, 
+	grabLocalTowns,
+	createConfirmCancelButtonRow,
+	handleCatchDelete
+} = require('../../uniHelperFunctions.js');
 const { createBasicPageButtons, loadRarStringMenu, loadPriceButts, handlePriceButtPicked, loadNameStringMenu, handleMatNameFilter, loadBasicBackButt } = require('./exported/tradeExtras.js');
-const { checkInboundTownMat, checkOutboundMat, checkInboundMat, checkOutboundTownMat } = require('../Development/Export/itemMoveContainer.js');
-const { updateUserCoins, spendUserCoins } = require('../Development/Export/uni_userPayouts.js');
-const { checkingRar, checkingRarID, baseCheckRarName } = require('../Development/Export/itemStringCore.js');
+// const { checkInboundTownMat, checkOutboundMat, checkInboundMat, checkOutboundTownMat } = require('../Development/Export/itemMoveContainer.js');
+// const { updateUserCoins, spendUserCoins } = require('../Development/Export/uni_userPayouts.js');
+// const { checkingRar, checkingRarID, baseCheckRarName } = require('../Development/Export/itemStringCore.js');
+const {
+    loadTownStoreButts,
+    loadClaimPlotButts,
+    loadOwnedPlotSelectButts,
+    loadCoreBuildTypeButtons,
+    loadCoreUpgradeTypeButtons,
+    createBuildTypeButtonRow,
+
+    generateTownDisplayEmbed,
+    generateLocalTownBonuses,
+    loadTownBuildingDisplayList,
+    loadTownCoreBuildingDisplayList,
+    handleTownMaterialStorageDisplay,
+    handleCoreCostDisplay,
+
+    grabTownCoreBuildings,
+    grabTownPlotList,
+    checkTownHasMaterials,
+    
+    handleJoinTown,
+    updateTownCanEditList,
+    updateTownMayor,
+    handleDepositIntoTown,
+    handleWithdrawFromTown,
+    handleTownPlotStatusUpdates,
+    handleTownPlotClaim,
+    handleBuildingOnTownPlot,
+    handleCoreBuildingConstruction
+} = require('./exported/townExtras.js');
+
 
 module.exports = {
 	data: new SlashCommandBuilder()
@@ -448,6 +491,11 @@ module.exports = {
 			}
 		};
 		const townRef = await grabTownFromOptions(subCom);
+
+		const countUserPlots = async (user, town, checkFor="Owned") => {
+			return (await grabTownPlotList(town, checkFor)).reduce((acc, plot) => {return acc + (plot.ownerid === user.userid);}, 0);
+		};
+
 		const mayorRef = await grabUser(townRef.mayorid) ?? 'None';
 
 		const localTowns = await grabLocalTowns(interaction.guild.id);
@@ -492,6 +540,11 @@ module.exports = {
 			embedPages: [],
 			pageFiles: [], // Files to display with embeds if used
 			pageDetails: [], // Details for each build/core in display list.
+			clergyQuest: {
+				active: false,
+				posIndex: 0,
+				callFunction: "" // REPLACED BY handleClergyMystQuest
+			},
 			costList: { // Material Cost Lists for core building constructs
 				need: [],
 				owned: []
@@ -501,106 +554,144 @@ module.exports = {
 			moveAmount: 0 // Amount selected for deposit/withdraw
 		};
 
+		// FUNCTION MOVED 
 		// DEPOSIT/WITHDRAW BUTTS
-		const matSelectButt = new ButtonBuilder()
-		.setCustomId('mat')
-		.setStyle(ButtonStyle.Primary)
-		.setLabel(`${subCom} Material`);
-		const coinSelectButt = new ButtonBuilder()
-		.setCustomId('coin')
-		.setStyle(ButtonStyle.Primary)
-		.setLabel(`${subCom} Coins`);
+		// /**
+		//  * This function loads the deposit/withdraw type select buttons
+		//  * @param {string} subCom Current Subcommand Name
+		//  * @returns {ButtonBuilder[]}
+		//  */
+		// function loadTownStoreButts(subCom){
+		// 	const matSelectButt = new ButtonBuilder()
+		// 	.setCustomId('mat')
+		// 	.setStyle(ButtonStyle.Primary)
+		// 	.setLabel(`${subCom} Material`);
+		// 	const coinSelectButt = new ButtonBuilder()
+		// 	.setCustomId('coin')
+		// 	.setStyle(ButtonStyle.Primary)
+		// 	.setLabel(`${subCom} Coins`);
 
-		const storeButtList = [matSelectButt, coinSelectButt];
+		// 	return [matSelectButt, coinSelectButt];
+		// }
+
+		const storeButtList = loadTownStoreButts(subCom);
 		const storeButtRow = new ActionRowBuilder().addComponents(storeButtList);
 
+		// FUNCTION MOVED 
 		// CLAIM PLOTS BUTTS
-		const openedPlotsButt = new ButtonBuilder()
-		.setCustomId('open-plot')
-		.setStyle(ButtonStyle.Primary)
-		.setLabel(`Claim Open Plot`);
-		const closedPlotsButt = new ButtonBuilder()
-		.setCustomId('closed-plot')
-		.setStyle(ButtonStyle.Primary)
-		.setDisabled(!permList.canEdit)
-		.setLabel(`Claim Closed Plot`);
+		// /**
+		//  * This function loads the claim town plot buttons
+		//  * @param {object} permList Users Town Permission Object 
+		//  * @returns {ButtonBuilder[]}
+		//  */
+		// function loadClaimPlotButts(permList){
+		// 	const openedPlotsButt = new ButtonBuilder()
+		// 	.setCustomId('open-plot')
+		// 	.setStyle(ButtonStyle.Primary)
+		// 	.setLabel(`Claim Open Plot`);
+		// 	const closedPlotsButt = new ButtonBuilder()
+		// 	.setCustomId('closed-plot')
+		// 	.setStyle(ButtonStyle.Primary)
+		// 	.setDisabled(!permList.canEdit)
+		// 	.setLabel(`Claim Closed Plot`);
 
-		const plotsButtList = [openedPlotsButt, closedPlotsButt];
+		// 	return [openedPlotsButt, closedPlotsButt];
+		// }
+
+		const plotsButtList = loadClaimPlotButts(permList);
 		const plotsButtRow = new ActionRowBuilder().addComponents(plotsButtList);
 
+		// FUNCTION MOVED 
 		// OWNED PLOTS BUTTS
-		const opOneButt = new ButtonBuilder()
-		.setCustomId('plot-one')
-		.setStyle(ButtonStyle.Primary)
-		.setLabel('Plot One');
-		const opTwoButt = new ButtonBuilder()
-		.setCustomId('plot-two')
-		.setStyle(ButtonStyle.Primary)
-		.setDisabled(!permList.canEdit)
-		.setLabel('Plot Two');
-		const opThreeButt = new ButtonBuilder()
-		.setCustomId('plot-three')
-		.setStyle(ButtonStyle.Primary)
-		.setDisabled(!permList.canEdit)
-		.setLabel('Plot Three');
+		// /**
+		//  * This function loads the owned plots selection buttons
+		//  * @param {object} permList Users Town Permission Object 
+		//  * @returns {ButtonBuilder[]}
+		//  */
+		// function loadOwnedPlotSelectButts(permList){
+		// 	const opOneButt = new ButtonBuilder()
+		// 	.setCustomId('plot-one')
+		// 	.setStyle(ButtonStyle.Primary)
+		// 	.setLabel('Plot One');
+		// 	const opTwoButt = new ButtonBuilder()
+		// 	.setCustomId('plot-two')
+		// 	.setStyle(ButtonStyle.Primary)
+		// 	.setDisabled(!permList.canEdit)
+		// 	.setLabel('Plot Two');
+		// 	const opThreeButt = new ButtonBuilder()
+		// 	.setCustomId('plot-three')
+		// 	.setStyle(ButtonStyle.Primary)
+		// 	.setDisabled(!permList.canEdit)
+		// 	.setLabel('Plot Three');
 
-		const oPlotsButtList = [opOneButt, opTwoButt, opThreeButt];
+		// 	return [opOneButt, opTwoButt, opThreeButt];
+		// }
+
+		const oPlotsButtList = loadOwnedPlotSelectButts(permList);
 		const oPlotsButtRow = new ActionRowBuilder().addComponents(oPlotsButtList);
 
+		// FUNCTION MOVED 
 		// CORE BUTT LIST
-		/**
-		 * This function loads the button list related to the given towns core buildings
-		 * 
-		 * Buttons are disabled if the ``Core Building`` **is** ``Built``
-		 * IDS: ``build-${coretype}``
-		 * @param {object} town Town DB Object
-		 * @returns {Promise <ButtonBuilder[]>}
-		 */
-		async function loadCoreBuildTypeButtons(town){
-			const townCores = await grabTownCoreBuildings(town);
+		// /**
+		//  * This function loads the button list related to the given towns core buildings
+		//  * 
+		//  * Buttons are disabled if the ``Core Building`` **is** ``Built``
+		//  * IDS: ``build-${coretype}``
+		//  * @param {object} town Town DB Object
+		//  * @returns {Promise <{noChoices: boolean, buttons: ButtonBuilder[]}>}
+		//  */
+		// async function loadCoreBuildTypeButtons(town){
+		// 	const townCores = await grabTownCoreBuildings(town);
 
-			const coreIsBuilt = (checkType) => townCores.some(core => core.build_type === checkType);
+		// 	let totalChoices = 5;
+		// 	const coreIsBuilt = (checkType) => townCores.some(core => core.build_type === checkType);
 
-			const coreTypes = ['grandhall', 'bank', 'market', 'tavern', 'clergy'];
-			const buttList = [];
-			for (const type of coreTypes){
-				const button = new ButtonBuilder()
-				.setCustomId(`build-${type}`)
-				.setStyle(ButtonStyle.Primary)
-				.setDisabled(coreIsBuilt(type))
-				.setLabel(`Build ${makeCapital(type)}`);
-				buttList.push(button);
-			}
+		// 	const coreTypes = ['grandhall', 'bank', 'market', 'tavern', 'clergy'];
+		// 	const buttList = [];
+		// 	for (const type of coreTypes){
+		// 		const isBuilt = coreIsBuilt(type);
+		// 		if (isBuilt) totalChoices--;
+		// 		const button = new ButtonBuilder()
+		// 		.setCustomId(`build-${type}`)
+		// 		.setStyle(ButtonStyle.Primary)
+		// 		.setDisabled(isBuilt)
+		// 		.setLabel(`Build ${makeCapital(type)}`);
+		// 		buttList.push(button);
+		// 	}
 
-			return buttList;
-		}
+		// 	return {noChoices: totalChoices === 0, buttons: buttList};
+		// }
 
-		/**
-		 * This function loads the button list related to the given towns core buildings
-		 * 
-		 * Buttons are disabled if the ``Core Building`` **is not** ``Built``
-		 * IDS: ``upgrade-${coretype}``
-		 * @param {object} town Town DB Object
-		 * @returns {Promise <ButtonBuilder[]>}
-		 */
-		async function loadCoreUpgradeTypeButtons(town){
-			const townCores = await grabTownCoreBuildings(town);
+		// FUNCTION MOVED 
+		// /**
+		//  * This function loads the button list related to the given towns core buildings
+		//  * 
+		//  * Buttons are disabled if the ``Core Building`` **is not** ``Built``
+		//  * IDS: ``upgrade-${coretype}``
+		//  * @param {object} town Town DB Object
+		//  * @returns {Promise <{noChoices: boolean, buttons: ButtonBuilder[]}>}
+		//  */
+		// async function loadCoreUpgradeTypeButtons(town){
+		// 	const townCores = await grabTownCoreBuildings(town);
 
-			const coreIsBuilt = (checkType) => townCores.some(core => core.build_type === checkType);
+		// 	let totalChoices = 5;
+		// 	const coreIsBuilt = (checkType) => townCores.some(core => core.build_type === checkType);
 
-			const coreTypes = ['grandhall', 'bank', 'market', 'tavern', 'clergy'];
-			const buttList = [];
-			for (const type of coreTypes){
-				const button = new ButtonBuilder()
-				.setCustomId(`upgrade-${type}`)
-				.setStyle(ButtonStyle.Primary)
-				.setDisabled(!coreIsBuilt(type))
-				.setLabel(`Upgrade ${makeCapital(type)}`);
-				buttList.push(button);
-			}
+		// 	const coreTypes = ['grandhall', 'bank', 'market', 'tavern', 'clergy'];
+		// 	const buttList = [];
+		// 	for (const type of coreTypes){
+		// 		const isBuilt = coreIsBuilt(type);
+		// 		if (!isBuilt) totalChoices--;
+		// 		const button = new ButtonBuilder()
+		// 		.setCustomId(`upgrade-${type}`)
+		// 		.setStyle(ButtonStyle.Primary)
+		// 		.setDisabled(!isBuilt)
+		// 		.setLabel(`Upgrade ${makeCapital(type)}`);
+		// 		buttList.push(button);
+		// 	}
 
-			return buttList;
-		}
+		// 	return {noChoices: totalChoices === 0, buttons: buttList};
+		// }
 
 
 		const rarSelectEmbed = new EmbedBuilder()
@@ -618,6 +709,7 @@ module.exports = {
 		let hasActionRow = true, usePagination = false, startNumSelect = false;
 		//useActionOneButt = false, useActionTwoButt = false;
 		let loadingBuildingList = false, buildLoadAnchor, loadingMatList = false, matLoadAnchor;
+		const userPlotCount = await countUserPlots(user, townRef, "Owned");
 		switch(subCom){
 			case "establish":
 			return await interaction.reply('This command is not yet available!!');
@@ -735,6 +827,7 @@ module.exports = {
 				startNumSelect = true;
 			break;
 			case "claimplot": // CLAIM PLOT MENU
+				if (!permList.canEdit && userPlotCount >= 1 || permList.canEdit && userPlotCount >= 3) return await interaction.reply({content: `You own ${userPlotCount} plot(s), which is the maximum amount you can own!! `, ephemeral: true});
 				// USE BUTTON NAV MENU
 				// PAGE 1: SELECT FROM
 				// - OPEN PLOTS
@@ -745,6 +838,9 @@ module.exports = {
 				// PAGE 2: CONFIRM
 			break;
 			case "buildplot": // BUILD ON OWNED PLOT MENU
+				if (!userPlotCount) return await interaction.reply({content: 'You do not own any plots yet! Use ``/town claimplot`` to obtain one!', ephemeral: true});
+				const emptyUserPlots = await countUserPlots(user, townRef, "Empty");
+				if (!emptyUserPlots) return await interaction.reply({content: 'You do not have any empty plots! View them using ``/town viewplot``', ephemeral: true});
 				// USE BUTTON NAV MENU
 				// PAGE 1: SELECT FROM
 				firstDisplayEmbed
@@ -753,6 +849,8 @@ module.exports = {
 				// PAGE 2: CONFIRM
 			break;
 			case "viewplot": // VIEW LOCAL TOWN BUILT ON PLOTS
+				const townBuiltPlots = await grabTownPlotList(townRef, "Built", true);
+				if (!townBuiltPlots) return await interaction.reply({content: 'This town does not have any buildings yet!', ephemeral: true});
 				// USE BUTTON NAV MENU
 				// PAGE 1: VIEW PAGES
 				actionOneButt
@@ -777,26 +875,32 @@ module.exports = {
 			break;
 			case "buildcore": // BUILD A TOWN CORE BUILDING
 				// USE BUTTON NAV MENU
+				const buildButtObj = await loadCoreBuildTypeButtons(townRef);
+				if (buildButtObj.noChoices) return await interaction.reply({content: 'This town has built all 5 core buildings! Use ``/town upgradecore`` instead!', ephemeral: true});
 				// PAGE 1: TYPE
 				firstDisplayEmbed
 				.setTitle(`== Build Core Building ==`)
 				.setDescription(`Select one of the options below.`);
-				buttonList.splice(0, 2, ...await loadCoreBuildTypeButtons(townRef));
+				buttonList.splice(0, 2, ...buildButtObj.buttons);
 				// PAGE 2: COST
 				// PAGE 3: CONFIRM
 			break;
 			case "upgradecore": // UPGRADE A TOWN CORE BUILDING
 				// USE BUTTON NAV MENU
+				const upgradeButtObj = await loadCoreUpgradeTypeButtons(townRef);
+				if (upgradeButtObj.noChoices) return await interaction.reply({content: 'This town has not built any core buildings! Use ``/town buildcore`` instead!', ephemeral: true});
 				// PAGE 1: TYPE
 				firstDisplayEmbed
 				.setTitle(`== Upgrade Core Building ==`)
 				.setDescription(`Select one of the options below.`);
-				buttonList.splice(0, 2, ...await loadCoreUpgradeTypeButtons(townRef));
+				buttonList.splice(0, 2, ...upgradeButtObj.buttons);
 				// PAGE 2: COST
 				// PAGE 3: CONFIRM
 			break;
 			case "viewcore": // VIEW BUILT TOWN CORE BUILDINGS
 				// USE BUTTON NAV MENU
+				const builtCoreCount = await grabTownCoreBuildings(townRef, true);
+				if (!builtCoreCount) return await interaction.reply({content: 'This town has not built any core buildings! Use ``/town buildcore`` to make one, or try looking in a different town!', ephemeral: true});
 				// PAGE 1: VIEW PAGES
 				actionOneButt
 				.setCustomId('view-core')
@@ -815,18 +919,26 @@ module.exports = {
 				navMenu.pageFiles = coreBuildViewObj.files;
 				navMenu.pageDetails = coreBuildViewObj.details;
 
+				if (coreBuildViewObj.clergyQuest.active){
+					navMenu.clergyQuest.active = true;
+					navMenu.clergyQuest.posIndex = coreBuildViewObj.clergyQuest.posIndex;
+					navMenu.clergyQuest.callFunction = coreBuildViewObj.clergyQuest.callFunction;
+				}
+
 				loadingBuildingList = true;
 				usePagination = true;
 			break;
 			case "storage": // VIEW TOWN OWNED MATERIALS
 				// USE BUTTON NAV MENU
+				const townStorageHasItem = await checkTownHasMaterials(townRef);
+				if (!townStorageHasItem) return await interaction.reply({content: 'This town has no stored materials! Use ``/town deposit`` to change that!', ephemeral: true});
 				// PAGE 1: VIEW PAGES
 				firstDisplayEmbed
 				.setTitle('== Loading Materials ==');
 
 				matLoadAnchor = await interaction.reply({embeds: [firstDisplayEmbed]});
 
-				const matViewObj = await handleTownMaterialStorageDisplay(townRef);
+				const matViewObj = await handleTownMaterialStorageDisplay(townRef, materialFiles);
 				navMenu.embedPages = matViewObj;
 				navMenu.lastPage = navMenu.embedPages.length - 1;
 
@@ -849,8 +961,6 @@ module.exports = {
 				replyObj.components = [actionButtRow.addComponents(buttonList)];
 			}
 		}
-
-		
 
 		const builtActionOneButt = actionOneButt.data?.custom_id !== undefined;
 		const builtActionTwoButt = actionTwoButt.data?.custom_id !== undefined;
@@ -1040,14 +1150,14 @@ module.exports = {
 						case "cancel-core": // BACK TO CORE TYPE SELECT
 							navMenu.typePicked = "";
 							if (subCom === 'buildcore'){
-								editWith = {embeds: [firstDisplayEmbed], components: [await loadCoreBuildTypeButtons(townRef)]};
-							} else editWith = {embeds: [firstDisplayEmbed], components: [await loadCoreUpgradeTypeButtons(townRef)]};
+								editWith = {embeds: [firstDisplayEmbed], components: replyObj.components};
+							} else editWith = {embeds: [firstDisplayEmbed], components: replyObj.components};
 						break;
 						case "back-core": // BACK TO CORE TYPE SELECT
 							navMenu.typePicked = "";
 							if (subCom === 'buildcore'){
-								editWith = {embeds: [firstDisplayEmbed], components: [await loadCoreBuildTypeButtons(townRef)]};
-							} else editWith = {embeds: [firstDisplayEmbed], components: [await loadCoreUpgradeTypeButtons(townRef)]};
+								editWith = {embeds: [firstDisplayEmbed], components: replyObj.components};
+							} else editWith = {embeds: [firstDisplayEmbed], components: replyObj.components};
 						break;
 					}
 				} else if (['mat', 'coin'].includes(c.customId)){ // STORE TYPE SELECT
@@ -1083,7 +1193,7 @@ module.exports = {
 					// Level ? Core Cost
 					// Handle Cost Display
 					navMenu.typePicked = c.customId.split('-')[1];
-					const coreCostOutcome = await handleCoreCostDisplay(townRef, navMenu, c.customId.split('-')[0]);
+					const coreCostOutcome = await handleCoreCostDisplay(townRef, navMenu, c.customId.split('-')[0], materialFiles);
 					navMenu.costList.need = coreCostOutcome.costObj.matCosts;
 					navMenu.costList.owned = coreCostOutcome.costObj.matsOwned;
 					const coreCostActionRow = (coreCostOutcome.canConfirm) ? [createConfirmCancelButtonRow('core')] : [loadBasicBackButt('core')];
@@ -1166,8 +1276,12 @@ module.exports = {
 						editWith = {embeds: [navMenu.embedPages[navMenu.curPage]], components: [pageButtRow], files: [navMenu.pageFiles[navMenu.curPage]]};
 					} else editWith = {embeds: [navMenu.embedPages[navMenu.curPage]], components: [pageButtRow]};
 					
-				} else if (['view-build', 'view-core']){
+				} else if (['view-build', 'view-core'].includes(c.customId)){
 					await c.followUp({embeds: [navMenu.pageDetails[navMenu.curPage]], ephemeral: true});
+
+					if (c.customId === 'view-core' && navMenu.curPage === navMenu.clergyQuest.posIndex){
+						await navMenu.clergyQuest.callFunction(user);
+					}
 
 					editWith = {embeds: [navMenu.embedPages[navMenu.curPage]], components: [pageButtRow], files: [navMenu.pageFiles[navMenu.curPage]]};
 				}
@@ -1255,756 +1369,800 @@ module.exports = {
 		// =====================
 
 
-		// Biome Background Texture Array:
-		// Used for Building canvas creation.
-		const biomeBTexList = ["None", "Forest", "Mountain", "Desert", "Plains", "Swamp", "Grassland"];
+		// // Biome Background Texture Array:
+		// // Used for Building canvas creation.
+		// const biomeBTexList = ["None", "Forest", "Mountain", "Desert", "Plains", "Swamp", "Grassland"];
 
+		// // FUNCTION MOVED
+		// /**
+		//  * This function checks if the given town has at least one stored material
+		//  * @param {object} town Town DB Object
+		//  * @returns {Promise <boolean>}
+		//  */
+		// async function checkTownHasMaterials(town){
+		// 	const hasMat = await TownMaterial.findOne({where: {townid: town.townid}});
+		// 	return !!hasMat;
+		// }
 
-		async function handleTownMaterialStorageDisplay(town){
-			const fullTownMaterialList = await TownMaterial.findAll({where: {townid: town.townid}})
-			const materialPages = [];
-			for (const [key, value] of materialFiles){
-				const matType = key;
-				const matEmbed = new EmbedBuilder()
-				.setTitle(`== ${makeCapital(matType)} Type Materials ==`);
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles loading the given towns owned material list display.
+		//  * @param {object} town Town DB Object
+		//  * @returns {Promise <EmbedBuilder[]>}
+		//  */
+		// async function handleTownMaterialStorageDisplay(town, materialFiles){
+		// 	const fullTownMaterialList = await TownMaterial.findAll({where: {townid: town.townid}})
+		// 	const materialPages = [];
+		// 	for (const [key, value] of materialFiles){
+		// 		const matType = key;
+		// 		const matEmbed = new EmbedBuilder()
+		// 		.setTitle(`== ${makeCapital(matType)} Type Materials ==`);
 
-				const matList = require(value);
-				matList.sort((a,b) => a.Rar_id - b.Rar_id);
+		// 		const matList = require(value);
+		// 		matList.sort((a,b) => a.Rar_id - b.Rar_id);
 
-				const refListLength = matList.length;
-				let missingAll = false;
+		// 		const refListLength = matList.length;
+		// 		let missingAll = false;
 
-				const matchingOwnedMats = (key === "unique") 
-				? fullTownMaterialList.filter(mat => mat.rar_id === 12)
-				: fullTownMaterialList.filter(mat => mat.mattype === key);
-				if (matchingOwnedMats.length === 0) missingAll = true;
+		// 		const matchingOwnedMats = (key === "unique") 
+		// 		? fullTownMaterialList.filter(mat => mat.rar_id === 12)
+		// 		: fullTownMaterialList.filter(mat => mat.mattype === key);
+		// 		if (matchingOwnedMats.length === 0) missingAll = true;
 
-				const orderedUMats = new Array(refListLength);
-				if (!missingAll){
-					let counter = 0;
-					for (const matRef of matList){
-						orderedUMats[counter] = (key === "unique") 
-						? matchingOwnedMats.filter(mat => mat.mattype === matRef.UniqueMatch)[0] ?? matRef.Rar_id
-						: matchingOwnedMats.filter(mat => mat.rar_id === matRef.Rar_id)[0] ?? matRef.Rar_id;
-						counter++;
-					}
-				}
+		// 		const orderedUMats = new Array(refListLength);
+		// 		if (!missingAll){
+		// 			let counter = 0;
+		// 			for (const matRef of matList){
+		// 				orderedUMats[counter] = (key === "unique") 
+		// 				? matchingOwnedMats.filter(mat => mat.mattype === matRef.UniqueMatch)[0] ?? matRef.Rar_id
+		// 				: matchingOwnedMats.filter(mat => mat.rar_id === matRef.Rar_id)[0] ?? matRef.Rar_id;
+		// 				counter++;
+		// 			}
+		// 		}
 
-				const matFields = await handleMaterialDisplay(orderedUMats, matList, missingAll);
-				matEmbed.addFields(matFields);
+		// 		const matFields = await handleMaterialDisplay(orderedUMats, matList, missingAll);
+		// 		matEmbed.addFields(matFields);
 				
-				materialPages.push(matEmbed);
-			}
+		// 		materialPages.push(matEmbed);
+		// 	}
 
-			return materialPages;
-		}
+		// 	return materialPages;
+		// }
 
-		/**
-         * This function constructs the field array to be used for displaying mats,
-         * by way of EmbedBuilder.addFields(finalFields)
-         * @param {(object|string)[]} matchMatList Array of owned materials
-         * @param {object[]} matRefList Material list array
-         * @param {boolean} emptyMatch true if no matching mats found
-         * @returns {promise <{name: string, value: string}[]>} Object array 
-         */
-        async function handleMaterialDisplay(matchMatList, matRefList, emptyMatch){
-            const finalFields = [];
+		// // FUNCTION MOVED
+		// /**
+        //  * This function constructs the field array to be used for displaying mats,
+        //  * by way of EmbedBuilder.addFields(finalFields)
+        //  * @param {(object|string)[]} matchMatList Array of owned materials
+        //  * @param {object[]} matRefList Material list array
+        //  * @param {boolean} emptyMatch true if no matching mats found
+        //  * @returns {promise <{name: string, value: string}[]>} Object array 
+        //  */
+        // async function handleMaterialDisplay(matchMatList, matRefList, emptyMatch){
+        //     const finalFields = [];
             
-            for (let i = 0; i < matRefList.length; i++){
-                let fieldName = '', fieldValue = '';
-                if (typeof matchMatList[i] === 'number' || emptyMatch){
-                    // Missing Material
-                    fieldName = `Unknown material of **${matRefList[i].Rarity}** rarity:`;
-                    fieldValue = 'Amount Owned: 0';
-                } else {
-                    // Matching Material
-                    fieldName = `~= ${matRefList[i].Rarity} Material =~`;
-                    fieldValue = `Name: **__${matchMatList[i].name}__**\nAmount Owned: **${matchMatList[i].amount}**`;
-                }
+        //     for (let i = 0; i < matRefList.length; i++){
+        //         let fieldName = '', fieldValue = '';
+        //         if (typeof matchMatList[i] === 'number' || emptyMatch){
+        //             // Missing Material
+        //             fieldName = `Unknown material of **${matRefList[i].Rarity}** rarity:`;
+        //             fieldValue = 'Amount Owned: 0';
+        //         } else {
+        //             // Matching Material
+        //             fieldName = `~= ${matRefList[i].Rarity} Material =~`;
+        //             fieldValue = `Name: **__${matchMatList[i].name}__**\nAmount Owned: **${matchMatList[i].amount}**`;
+        //         }
 
-                finalFields.push({name: fieldName, value: fieldValue});
-            }
+        //         finalFields.push({name: fieldName, value: fieldValue});
+        //     }
 
-            return finalFields;
-        }
+        //     return finalFields;
+        // }
 
-		/**
-		 * This function handles both newly built core buildings and upgrading existing ones.
-		 * 
-		 * Is responsable for returning the core building display upon completion.
-		 * @param {object} town 
-		 * @param {object} navMenu 
-		 * @returns {Promise <{embeds: EmbedBuilder[], files: AttachmentBuilder[]}>}
-		 */
-		async function handleCoreBuildingConstruction(town, navMenu){
-			let coreBuilding = await CoreBuilding.findOrCreate({
-				where: {
-					townid: town.townid,
-					build_type: navMenu.typePicked
-				},
-				defaults: {
-					level: 1,
-					build_status: "Level 1",
-					background_tex: biomeBTexList.indexOf(town.local_biome.split('-')[0])
-				}
-			});
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles both newly built core buildings and upgrading existing ones.
+		//  * 
+		//  * Is responsable for returning the core building display upon completion.
+		//  * @param {object} town Town DB Object
+		//  * @param {object} navMenu Nav menu object holding all user selected values needed
+		//  * @returns {Promise <{embeds: EmbedBuilder[], files: AttachmentBuilder[]}>}
+		//  */
+		// async function handleCoreBuildingConstruction(town, navMenu){
+		// 	let coreBuilding = await CoreBuilding.findOrCreate({
+		// 		where: {
+		// 			townid: town.townid,
+		// 			build_type: navMenu.typePicked
+		// 		},
+		// 		defaults: {
+		// 			level: 1,
+		// 			build_status: "Level 1",
+		// 			background_tex: biomeBTexList.indexOf(town.local_biome.split('-')[0])
+		// 		}
+		// 	});
 
-			// Was created
-			let wasUpgraded = true;
-			if (coreBuilding[1]){
-				wasUpgraded = false;
-				await coreBuilding[0].save().then(async cb => {return await cb.reload()});
-			} else {
-				await coreBuilding[0].update({level: coreBuilding[0].level + 1, build_status: `Level ${coreBuilding[0].level + 1}`})
-				.then(async cb => await cb.save())
-				.then(async cb => {return await cb.reload()});
-			}
+		// 	// Was created
+		// 	let wasUpgraded = true;
+		// 	if (coreBuilding[1]){
+		// 		wasUpgraded = false;
+		// 		await coreBuilding[0].save().then(async cb => {return await cb.reload()});
+		// 	} else {
+		// 		await coreBuilding[0].update({level: coreBuilding[0].level + 1, build_status: `Level ${coreBuilding[0].level + 1}`})
+		// 		.then(async cb => await cb.save())
+		// 		.then(async cb => {return await cb.reload()});
+		// 	}
 
-			coreBuilding = coreBuilding[0];
+		// 	coreBuilding = coreBuilding[0];
 
-			await town.update({[`${navMenu.typePicked}_status`]: `Built: Level ${coreBuilding.level}`})
-			.then(async t => await t.save())
-			.then(async t => {return await t.reload()});
+		// 	await town.update({[`${navMenu.typePicked}_status`]: `Built: Level ${coreBuilding.level}`})
+		// 	.then(async t => await t.save())
+		// 	.then(async t => {return await t.reload()});
 
-			const coreBuildingPngFile = await loadBuilding(coreBuilding);
+		// 	const coreBuildingPngFile = await loadBuilding(coreBuilding);
 
-			const dynBuildDesc = (wasUpgraded) ? `${makeCapital(navMenu.typePicked)} upgraded to level **${coreBuilding.level}**!`: `${makeCapital(navMenu.typePicked)} built at level 1!`;
-			const constructType = (wasUpgraded) ? 'Upgraded': 'Built';
-			const coreBuildingCreateEmbed = new EmbedBuilder()
-			.setTitle(`== Core Building ${constructType} ==`)
-			.setColor('DarkGold')
-			.setDescription(dynBuildDesc);
+		// 	const dynBuildDesc = (wasUpgraded) ? `${makeCapital(navMenu.typePicked)} upgraded to level **${coreBuilding.level}**!`: `${makeCapital(navMenu.typePicked)} built at level 1!`;
+		// 	const constructType = (wasUpgraded) ? 'Upgraded': 'Built';
+		// 	const coreBuildingCreateEmbed = new EmbedBuilder()
+		// 	.setTitle(`== Core Building ${constructType} ==`)
+		// 	.setColor('DarkGold')
+		// 	.setDescription(dynBuildDesc);
 
-			return {embeds: [coreBuildingCreateEmbed], files: [coreBuildingPngFile]};
-		}
+		// 	return {embeds: [coreBuildingCreateEmbed], files: [coreBuildingPngFile]};
+		// }
 
-		/**
-		 * This function handles building the display for the upgrade costs given the construction type picked
-		 * @param {object} town Town DB Object
-		 * @param {object} navMenu Nav menu object holding all user selected values needed
-		 * @param {string} constructType One of ``build`` | ``upgrade``
-		 * @returns {Promise<{canConfirm: boolean, embed: EmbedBuilder, costObj: {matCosts: {name: string, matType: string, amount: number, rar_id: number}[], matsOwned: {name: string, matType: string, amount: number, rar_id: number}[]}>}
-		 */
-		async function handleCoreCostDisplay(town, navMenu, constructType){
-			// Town is needed for checking owned material amounts
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles building the display for the upgrade costs given the construction type picked
+		//  * @param {object} town Town DB Object
+		//  * @param {object} navMenu Nav menu object holding all user selected values needed
+		//  * @param {string} constructType One of ``build`` | ``upgrade``
+		//  * @returns {Promise<{canConfirm: boolean, embed: EmbedBuilder, costObj: {matCosts: {name: string, matType: string, amount: number, rar_id: number}[], matsOwned: {name: string, matType: string, amount: number, rar_id: number}[]}}>}
+		//  */
+		// async function handleCoreCostDisplay(town, navMenu, constructType){
+		// 	// Town is needed for checking owned material amounts
 
-			let coreBuildCostObj;
-			switch(constructType){
-				case "build":
-					coreBuildCostObj = await handleCoreCostCalc(town, 0, navMenu.typePicked);
-				break;
-				case "upgrade":
-					const coreRef = (await grabTownCoreBuildings(town)).find(core => core.build_type === navMenu.typePicked);
-					coreBuildCostObj = await handleCoreCostCalc(town, coreRef.level, coreRef.build_type);
-				break;
-			}
+		// 	let coreBuildCostObj;
+		// 	switch(constructType){
+		// 		case "build":
+		// 			coreBuildCostObj = await handleCoreCostCalc(town, 0, navMenu.typePicked);
+		// 		break;
+		// 		case "upgrade":
+		// 			const coreRef = (await grabTownCoreBuildings(town)).find(core => core.build_type === navMenu.typePicked);
+		// 			coreBuildCostObj = await handleCoreCostCalc(town, coreRef.level, coreRef.build_type);
+		// 		break;
+		// 	}
 
-			const costEmbed = new EmbedBuilder()
-			.setTitle(`== ${makeCapital(constructType)} Cost ==`)
-			.setDescription(`Material Requirements to begin core building construction:`);
+		// 	const costEmbed = new EmbedBuilder()
+		// 	.setTitle(`== ${makeCapital(constructType)} Cost ==`)
+		// 	.setDescription(`Material Requirements to begin core building construction:`);
 
-			const finalFields = [];
-			let idxMatch = 0, canConfirm = true;
-			for (const matNeeded of coreBuildCostObj.matCosts){
-				const matOwned = coreBuildCostObj.matsOwned[idxMatch];
-				if (canConfirm && matOwned.amount < matNeeded.amount) canConfirm = false;
-				finalFields.push({name: `== ${matNeeded.name} ==`, value: `Rarity: ${baseCheckRarName(matNeeded.rar_id)}\n(Owned/Needed) Amount: ${matOwned.amount}/${matNeeded.amount}`});
-				idxMatch++;
-			}
+		// 	const finalFields = [];
+		// 	let idxMatch = 0, canConfirm = true;
+		// 	for (const matNeeded of coreBuildCostObj.matCosts){
+		// 		const matOwned = coreBuildCostObj.matsOwned[idxMatch];
+		// 		if (canConfirm && matOwned.amount < matNeeded.amount) canConfirm = false;
+		// 		finalFields.push({name: `== ${matNeeded.name} ==`, value: `Rarity: ${baseCheckRarName(matNeeded.rar_id)}\n(Owned/Needed) Amount: ${matOwned.amount}/${matNeeded.amount}`});
+		// 		idxMatch++;
+		// 	}
 
-			costEmbed.addFields(finalFields);
+		// 	costEmbed.addFields(finalFields);
 
-			return {canConfirm: canConfirm, embed: costEmbed, costObj: coreBuildCostObj};
-		}
+		// 	return {canConfirm: canConfirm, embed: costEmbed, costObj: coreBuildCostObj};
+		// }
 
-		/**
-		 * This function handles calculating the cost to build/upgrade the given ``coreType`` building.
-		 * @param {object} town Town DB Object
-		 * @param {number} level Current Level of the given core building
-		 * @param {string} coreType Core Building ``build_type``
-		 * @returns {Promise <{matCosts: {name: string, matType: string, amount: number, rar_id: number}[], matsOwned: {name: string, matType: string, amount: number, rar_id: number}[]}>}
-		 */
-		async function handleCoreCostCalc(town, level, coreType){
-			const matTypeConverter = new Map([
-				['rocky', 'Stone'],
-				['woody', 'Wood'],
-				['metalic', 'Metal'],
-				['skinny', 'Hide'],
-				['slimy', 'Slime']
-			]);
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles calculating the cost to build/upgrade the given ``coreType`` building.
+		//  * @param {object} town Town DB Object
+		//  * @param {number} level Current Level of the given core building
+		//  * @param {string} coreType Core Building ``build_type``
+		//  * @returns {Promise <{matCosts: {name: string, matType: string, amount: number, rar_id: number}[], matsOwned: {name: string, matType: string, amount: number, rar_id: number}[]}>}
+		//  */
+		// async function handleCoreCostCalc(town, level, coreType){
+		// 	const matTypeConverter = new Map([
+		// 		['rocky', 'Stone'],
+		// 		['woody', 'Wood'],
+		// 		['metalic', 'Metal'],
+		// 		['skinny', 'Hide'],
+		// 		['slimy', 'Slime']
+		// 	]);
 
-			const coreCostRef = coreBuildCostList.find(core => core.Name === coreType);
+		// 	const coreCostRef = coreBuildCostList.find(core => core.Name === coreType);
 
-			const handleCoreMaterialScale = (lvl, costObj) => {
-				return costObj.Amount * 1 + (lvl - (costObj.Level));
-			};
+		// 	const handleCoreMaterialScale = (lvl, costObj) => {
+		// 		return costObj.Amount * (1 + (lvl - (costObj.Level)));
+		// 	};
 
-			const handleCoreMaterialRarity = (lvl, costObj) => {
-				return costObj.Rarity + Math.floor(lvl / 2);
-			};
+		// 	const handleCoreMaterialRarity = (lvl, costObj) => {
+		// 		return costObj.Rarity + Math.floor(lvl / 2);
+		// 	};
 
-			const nextLevel = level + 1;
-			const materialCostList = [], materialOwnedList = [];
-			for (const [matType, convType] of matTypeConverter){
-				// Check if next level !requires convType
-				if (coreCostRef[convType].Level > nextLevel) continue;
-				const curMatCostRef = coreCostRef[convType];
+		// 	const nextLevel = level + 1;
+		// 	const materialCostList = [], materialOwnedList = [];
+		// 	for (const [matType, convType] of matTypeConverter){
+		// 		// Check if next level !requires convType
+		// 		if (coreCostRef[convType].Level > nextLevel) continue;
+		// 		const curMatCostRef = coreCostRef[convType];
 
-				const matNumNeeded = handleCoreMaterialScale(nextLevel, curMatCostRef);
-				const matRarNeeded = handleCoreMaterialRarity(nextLevel, curMatCostRef);
+		// 		const matNumNeeded = handleCoreMaterialScale(nextLevel, curMatCostRef);
+		// 		const matRarNeeded = handleCoreMaterialRarity(nextLevel, curMatCostRef);
 				
-				const matRef = grabMaterialByRarID(matRarNeeded, matType, materialFiles);
+		// 		const matRef = grabMaterialByRarID(matRarNeeded, matType, materialFiles);
 
-				materialCostList.push({name: matRef.Name, matType: matType, amount: matNumNeeded, rar_id: matRarNeeded});
+		// 		materialCostList.push({name: matRef.Name, matType: matType, amount: matNumNeeded, rar_id: matRarNeeded});
 			
-				const townOwnsMat = await TownMaterial.findOne({where: {townid: town.townid, name: matRef.Name}});
-				materialOwnedList.push({name: matRef.Name, matType: matType, amount: townOwnsMat.amount ?? 0, rar_id: matRarNeeded});
-			}
+		// 		const townOwnsMat = await TownMaterial.findOne({where: {townid: town.townid, name: matRef.Name}});
+		// 		materialOwnedList.push({name: matRef.Name, matType: matType, amount: townOwnsMat?.amount ?? 0, rar_id: matRarNeeded});
+		// 	}
 
-			return {matCosts: materialCostList, matsOwned: materialOwnedList};
-		}
+		// 	return {matCosts: materialCostList, matsOwned: materialOwnedList};
+		// }
 
+		// // FUNCTION MOVED
+		// /**
+		//  * This function retrieves a material ref object based on the given ``rar`` & ``matType``
+		//  * @param {number} rar RarID to search for
+		//  * @param {string} matType Material Type to search for
+		//  * @param {Collection} materialFiles Discord Collection object
+		//  * @returns {object}
+		//  */
+		// function grabMaterialByRarID(rar, matType, materialFiles){
+		// 	const matFile = require(materialFiles.get(matType));
 
-		/**
-		 * This function retrieves a material ref object based on the given ``rar`` & ``matType``
-		 * @param {number} rar RarID to search for
-		 * @param {string} matType Material Type to search for
-		 * @param {Collection} materialFiles Discord Collection object
-		 * @returns {object}
-		 */
-		function grabMaterialByRarID(rar, matType, materialFiles){
-			const matFile = require(materialFiles.get(matType));
+		// 	const matRef = matFile.find(mat => mat.Rar_id === rar);
 
-			const matRef = matFile.find(mat => mat.Rar_id === rar);
+		// 	return matRef;
+		// }
 
-			return matRef;
-		}
+		// // FUNCTION MOVED
+		// /**
+		//  * This function loads all player made buildings located within the given town.
+		//  * @param {object} town Town DB Object
+		//  * @returns {Promise <{embeds: EmbedBuilder[], files: AttachmentBuilder[], details: EmbedBuilder[]}>}
+		//  */
+		// async function loadTownBuildingDisplayList(town){
+		// 	const returnObj = {embeds: [], files: [], details: []};
 
-		/**
-		 * This function loads all player made buildings located within the given town.
-		 * @param {object} town Town DB Object
-		 * @returns {Promise <{embeds: EmbedBuilder[], files: AttachmentBuilder[], details: EmbedBuilder[]}>}
-		 */
-		async function loadTownBuildingDisplayList(town){
-			const returnObj = {embeds: [], files: [], details: []};
+		// 	const builtPlots = await grabTownPlotList(town, "Built");
 
-			const builtPlots = await grabTownPlotList(town, "Built");
+		// 	for (const plot of builtPlots){
+		// 		const building = await PlayerBuilding.findOne({where: {plotid: plot.plotid}});
 
-			for (const plot of builtPlots){
-				const building = await PlayerBuilding.findOne({where: {plotid: plot.plotid}});
+		// 		const mainEmbed = new EmbedBuilder()
+		// 		.setTitle(`== ${makeCapital(building.build_type)} ==`)
+		// 		.setColor('DarkGold');
+		// 		returnObj.embeds.push(mainEmbed);
 
-				const mainEmbed = new EmbedBuilder()
-				.setTitle(`== ${makeCapital(building.build_type)} ==`)
-				.setColor('DarkGold');
-				returnObj.embeds.push(mainEmbed);
+		// 		const mainFile = await loadBuilding(building);
+		// 		returnObj.files.push(mainFile);
 
-				const mainFile = await loadBuilding(building);
-				returnObj.files.push(mainFile);
+		// 		const detailEmbed = new EmbedBuilder()
+		// 		.setTitle('== Building Details ==')
+		// 		.setColor('DarkNavy')
+		// 		.setDescription(`Town: ${makeCapital(town.name)}\nOwner: ${makeCapital((await grabUser(building.ownerid)).username)}\nBand: No Linked Band`);
+		// 		returnObj.details.push(detailEmbed);
+		// 	}
 
-				const detailEmbed = new EmbedBuilder()
-				.setTitle('== Building Details ==')
-				.setColor('DarkNavy')
-				.setDescription(`Town: ${makeCapital(town.name)}\nOwner: ${makeCapital((await grabUser(building.ownerid)).username)}\nBand: No Linked Band`);
-				returnObj.details.push(detailEmbed);
-			}
+		// 	return returnObj;
+		// }
 
-			return returnObj;
-		}
+		// // FUNCTION MOVED
+		// /**
+		//  * This function loads all built core buildings for the given town, ready to be displayed
+		//  * @param {object} town Town DB Object
+		//  * @returns {Promise <{embeds: EmbedBuilder[], files: AttachmentBuilder[], details: EmbedBuilder[], clergyQuest: {active: boolean, posIndex: number, callFunction: handleClergyMystQuest}}>}
+		//  */
+		// async function loadTownCoreBuildingDisplayList(town){
+		// 	const returnObj = {embeds: [], files: [], details: [], clergyQuest: {active: false, posIndex: 0, callFunction: handleClergyMystQuest}};
+		// 	const townCores = await grabTownCoreBuildings(town);
 
-		/**
-		 * This function loads all built core buildings for the given town, ready to be displayed
-		 * @param {object} town Town DB Object
-		 * @returns {Promise <{embeds: EmbedBuilder[], files: AttachmentBuilder[], details: EmbedBuilder[]}>}
-		 */
-		async function loadTownCoreBuildingDisplayList(town){
-			const returnObj = {embeds: [], files: [], details: []};
-			const townCores = await grabTownCoreBuildings(town);
+		// 	for (const core of townCores){
+		// 		const mainEmbed = new EmbedBuilder()
+		// 		.setTitle(`== ${makeCapital(core.build_type)} ==`)
+		// 		.setColor('DarkGold');
+		// 		returnObj.embeds.push(mainEmbed);
 
-			for (const core of townCores){
-				const mainEmbed = new EmbedBuilder()
-				.setTitle(`== ${makeCapital(core.build_type)} ==`)
-				.setColor('DarkGold');
-				returnObj.embeds.push(mainEmbed);
+		// 		const mainFile = await loadBuilding(core);
+		// 		returnObj.files.push(mainFile);
 
-				const mainFile = await loadBuilding(core);
-				returnObj.files.push(mainFile);
+		// 		const detailEmbed = new EmbedBuilder()
+		// 		.setTitle('== Core Building Details ==')
+		// 		.setColor('DarkNavy')
+		// 		.setDescription(`Town: ${makeCapital(town.name)}\n${makeCapital(core.build_type)} Level: ${core.level}`);
+		// 		returnObj.details.push(detailEmbed);
 
-				const detailEmbed = new EmbedBuilder()
-				.setTitle('== Core Building Details ==')
-				.setColor('DarkNavy')
-				.setDescription(`Town: ${makeCapital(town.name)}\n${makeCapital(core.build_type)} Level: ${core.level}`);
-				returnObj.details.push(detailEmbed);
-			}
+		// 		if (core.build_type === 'clergy'){
+		// 			returnObj.clergyQuest.active = true;
+		// 			returnObj.clergyQuest.posIndex = returnObj.details.length - 1;
+		// 		}
+		// 	}
 
-			return returnObj;
-		}
+		// 	return returnObj;
+		// }
 
-		/**
-		 * This function handles building the selected build type on on of the users owned && empty plots.
-		 * 
-		 * Constructs the building display image for the newly created building and returns it.
-		 * @param {object} town Town DB Object
-		 * @param {object} user UserData DB Object, user building
-		 * @param {object} navMenu Nav menu object holding all user selected values needed
-		 * @returns {Promise<{embeds: EmbedBuilder[]; files: (string | AttachmentBuilder)[];}>}
-		 */
-		async function handleBuildingOnTownPlot(town, user, navMenu){
-			const townOwnedPlots = await grabTownPlotList(town, "Empty");
-			const ownedPlotPicked = townOwnedPlots.find(plot => plot.ownerid === user.userid);  // .filter(plot => plot.ownerid === user.userid)[0];
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles building the selected build type on on of the users owned && empty plots.
+		//  * 
+		//  * Constructs the building display image for the newly created building and returns it.
+		//  * @param {object} town Town DB Object
+		//  * @param {object} user UserData DB Object, user building
+		//  * @param {object} navMenu Nav menu object holding all user selected values needed
+		//  * @returns {Promise<{embeds: EmbedBuilder[]; files: (string | AttachmentBuilder)[];}>}
+		//  */
+		// async function handleBuildingOnTownPlot(town, user, navMenu){
+		// 	const townOwnedPlots = await grabTownPlotList(town, "Empty");
+		// 	const ownedPlotPicked = townOwnedPlots.find(plot => plot.ownerid === user.userid);  // .filter(plot => plot.ownerid === user.userid)[0];
 
-			await ownedPlotPicked.update({empty: false})
-			.then(async tp => await tp.save())
-			.then(async tp => {return await tp.reload()});
+		// 	await ownedPlotPicked.update({empty: false})
+		// 	.then(async tp => await tp.save())
+		// 	.then(async tp => {return await tp.reload()});
 
-			await handleUpdateTownPlotTypeAmounts(town);
+		// 	await handleUpdateTownPlotTypeAmounts(town);
 
-			const building = await PlayerBuilding.create({
-				townid: town.townid,
-				ownerid: user.userid,
-				plotid: ownedPlotPicked.plotid,
-				can_edit: user.userid,
-				build_type: navMenu.typePicked,
-				background_tex: biomeBTexList.indexOf(town.local_biome.split('-')[0])
-			}).then(async b => await b.save()).then(async b => {return await b.reload()});
+		// 	const building = await PlayerBuilding.create({
+		// 		townid: town.townid,
+		// 		ownerid: user.userid,
+		// 		plotid: ownedPlotPicked.plotid,
+		// 		can_edit: user.userid,
+		// 		build_type: navMenu.typePicked,
+		// 		background_tex: biomeBTexList.indexOf(town.local_biome.split('-')[0])
+		// 	}).then(async b => await b.save()).then(async b => {return await b.reload()});
 		
-			const buildingPngFile = await loadBuilding(building);
+		// 	const buildingPngFile = await loadBuilding(building);
 
-			let dynBuildDesc = `New ${makeCapital(navMenu.typePicked)} created!`;
-			if (navMenu.typePicked === 'house') dynBuildDesc += `\nWelcome home! <3`;
-			const buildingCreateEmbed = new EmbedBuilder()
-			.setTitle('== Building Created ==')
-			.setColor('DarkGold')
-			.setDescription(dynBuildDesc);
+		// 	let dynBuildDesc = `New ${makeCapital(navMenu.typePicked)} created!`;
+		// 	if (navMenu.typePicked === 'house') dynBuildDesc += `\nWelcome home! <3`;
+		// 	const buildingCreateEmbed = new EmbedBuilder()
+		// 	.setTitle('== Building Created ==')
+		// 	.setColor('DarkGold')
+		// 	.setDescription(dynBuildDesc);
 
-			return {embeds: [buildingCreateEmbed], files: [buildingPngFile]};
-		}
+		// 	return {embeds: [buildingCreateEmbed], files: [buildingPngFile]};
+		// }
 
-		/**
-		 * This function loads all plot building option buttons
-		 * 
-		 * houseButt: ``build-house``
-		 * @returns {ActionRowBuilder}
-		 */
-		function createBuildTypeButtonRow(){
-			const houseButt = new ButtonBuilder()
-			.setCustomId('build-house')
-			.setStyle(ButtonStyle.Primary)
-			.setLabel('Build House!');
+		// // FUNCTION MOVED
+		// /**
+		//  * This function loads all plot building option buttons
+		//  * 
+		//  * houseButt: ``build-house``
+		//  * @returns {ActionRowBuilder}
+		//  */
+		// function createBuildTypeButtonRow(){
+		// 	const houseButt = new ButtonBuilder()
+		// 	.setCustomId('build-house')
+		// 	.setStyle(ButtonStyle.Primary)
+		// 	.setLabel('Build House!');
 
-			const buttonRow = new ActionRowBuilder().addComponents(houseButt);
+		// 	const buttonRow = new ActionRowBuilder().addComponents(houseButt);
 
-			return buttonRow;
-		}
+		// 	return buttonRow;
+		// }
 
-		/**
-		 * This function handles claiming a random unowned plot, assigning the plot to the given user,
-		 * then updates the plot counts of the given town.
-		 * @param {object} town Town DB Object
-		 * @param {object} user UserData DB Object, user claiming plot
-		 * @param {object} navMenu Nav menu object holding all user selected values needed
-		 */
-		async function handleTownPlotClaim(town, user, navMenu){
-			const townPlots = await grabTownPlotList(town, navMenu.navType);
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles claiming a random unowned plot, assigning the plot to the given user,
+		//  * then updates the plot counts of the given town.
+		//  * @param {object} town Town DB Object
+		//  * @param {object} user UserData DB Object, user claiming plot
+		//  * @param {object} navMenu Nav menu object holding all user selected values needed
+		//  */
+		// async function handleTownPlotClaim(town, user, navMenu){
+		// 	const townPlots = await grabTownPlotList(town, navMenu.navType);
 
-			const thePlot = randArrPos(townPlots);
+		// 	const thePlot = randArrPos(townPlots);
 
-			await thePlot.update({ownerid: user.userid, private: true})
-			.then(async tp => await tp.save())
-			.then(async tp => {return await tp.reload()});
+		// 	await thePlot.update({ownerid: user.userid, private: true})
+		// 	.then(async tp => await tp.save())
+		// 	.then(async tp => {return await tp.reload()});
 
-			await handleUpdateTownPlotTypeAmounts(town);
-		}
+		// 	await handleUpdateTownPlotTypeAmounts(town);
+		// }
 
-		/**
-		 * This function handles updating ``TownPlot.private`` for the given towns plots of ``currentPlotStatus``. 
-		 * @param {object} town Town DB Object
-		 * @param {object} navMenu Nav menu object holding all user selected values needed
-		 */
-		async function handleTownPlotStatusUpdates(town, navMenu){
-			const grabType = (navMenu.typePicked === "Open") ? "Closed" : "Open";
-			const townPlots = await grabTownPlotList(town, grabType);
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles updating ``TownPlot.private`` for the given towns plots of ``currentPlotStatus``. 
+		//  * @param {object} town Town DB Object
+		//  * @param {object} navMenu Nav menu object holding all user selected values needed
+		//  */
+		// async function handleTownPlotStatusUpdates(town, navMenu){
+		// 	const grabType = (navMenu.typePicked === "Open") ? "Closed" : "Open";
+		// 	const townPlots = await grabTownPlotList(town, grabType);
 
-			console.log('Total Plots: %d', townPlots.length);
+		// 	console.log('Total Plots: %d', townPlots.length);
 
-			for (let i = 0; i < navMenu.moveAmount; i++){
-				await townPlots[i].update({private: navMenu.typePicked !== 'Open'})
-				.then(async tp => await tp.save())
-				.then(async tp => {return await tp.reload()});
-			}
+		// 	for (let i = 0; i < navMenu.moveAmount; i++){
+		// 		await townPlots[i].update({private: navMenu.typePicked !== 'Open'})
+		// 		.then(async tp => await tp.save())
+		// 		.then(async tp => {return await tp.reload()});
+		// 	}
 
-			await handleUpdateTownPlotTypeAmounts(town);
-		}
+		// 	await handleUpdateTownPlotTypeAmounts(town);
+		// }
 
-		/**
-		 * This function checks and updates any plots !matching amount values found.
-		 * @param {object} town Town DB Object
-		 */
-		async function handleUpdateTownPlotTypeAmounts(town){
-			const openPlots = await grabTownPlotList(town, "Open", true);
-			if (town.openplots !== openPlots) await town.update({openplots: openPlots}).then(async t => await t.save()).then(async t => {return await t.reload()});
-			const closedPlots = await grabTownPlotList(town, "Closed", true);
-			if (town.closedplots !== closedPlots) await town.update({closedplots: closedPlots}).then(async t => await t.save()).then(async t => {return await t.reload()});
-			const ownedPlots = await grabTownPlotList(town, "Owned", true);
-			if (town.ownedplots !== ownedPlots) await town.update({ownedplots: ownedPlots}).then(async t => await t.save()).then(async t => {return await t.reload()});
-			const builtPlots = await grabTownPlotList(town, "Built", true);
-			if (town.buildcount !== builtPlots) await town.update({buildcount: builtPlots}).then(async t => await t.save()).then(async t => {return await t.reload()});
+		// // FUNCTION MOVED
+		// /**
+		//  * This function checks and updates any plots !matching amount values found.
+		//  * @param {object} town Town DB Object
+		//  */
+		// async function handleUpdateTownPlotTypeAmounts(town){
+		// 	const openPlots = await grabTownPlotList(town, "Open", true);
+		// 	if (town.openplots !== openPlots) await town.update({openplots: openPlots}).then(async t => await t.save()).then(async t => {return await t.reload()});
+		// 	const closedPlots = await grabTownPlotList(town, "Closed", true);
+		// 	if (town.closedplots !== closedPlots) await town.update({closedplots: closedPlots}).then(async t => await t.save()).then(async t => {return await t.reload()});
+		// 	const ownedPlots = await grabTownPlotList(town, "Owned", true);
+		// 	if (town.ownedplots !== ownedPlots) await town.update({ownedplots: ownedPlots}).then(async t => await t.save()).then(async t => {return await t.reload()});
+		// 	const builtPlots = await grabTownPlotList(town, "Built", true);
+		// 	if (town.buildcount !== builtPlots) await town.update({buildcount: builtPlots}).then(async t => await t.save()).then(async t => {return await t.reload()});
 
-		}
+		// }
 
-		/**
-		 * This function handles locating all plots belonging to the given town.
-		 * 
-		 * It then filters them based on the provided ``plotStatus``.
-		 * @param {object} town Town DB Object
-		 * @param {string} plotStatus One of: ``Owned``, ``Closed``, ``Open``, ``Built``, ``Empty``
-		 * @param {boolean} returnLength Default: ``false``, set to ``true`` to return found plots list length
-		 * @returns {Promise <object[] | number>}
-		 */
-		async function grabTownPlotList(town, plotStatus, returnLength=false){
-			const townPlots = await TownPlots.findAll({where: {townid: town.townid}});
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles locating all plots belonging to the given town.
+		//  * 
+		//  * It then filters them based on the provided ``plotStatus``.
+		//  * @param {object} town Town DB Object
+		//  * @param {string} plotStatus One of: ``Owned``, ``Closed``, ``Open``, ``Built``, ``Empty``
+		//  * @param {boolean} returnLength Default: ``false``, set to ``true`` to return found plots list length
+		//  * @returns {Promise <object[] | number>}
+		//  */
+		// async function grabTownPlotList(town, plotStatus, returnLength=false){
+		// 	const townPlots = await TownPlots.findAll({where: {townid: town.townid}});
 
-			const hasOwner = (plot) => plot.ownerid !== '0';
-			const hasBuild = (plot) => !plot.empty;
-			const isClosed = (plot) => plot.private;
+		// 	const hasOwner = (plot) => plot.ownerid !== '0';
+		// 	const hasBuild = (plot) => !plot.empty;
+		// 	const isClosed = (plot) => plot.private;
 
-			let filteredPlotList = [];
-			switch(plotStatus){
-				case "Owned":
-					filteredPlotList = townPlots.filter(plot => hasOwner(plot));
-				break;
-				case "Closed":
-					filteredPlotList = townPlots.filter(plot => isClosed(plot) && !hasOwner(plot));
-				break;
-				case "Open":
-					filteredPlotList = townPlots.filter(plot => !isClosed(plot) && !hasOwner(plot));
-				break;
-				case "Built":
-					filteredPlotList = townPlots.filter(plot => hasBuild(plot));
-				break;
-				case "Empty":
-					filteredPlotList = townPlots.filter(plot => !hasBuild(plot) && hasOwner(plot));
-				break;
-				default:
-					filteredPlotList = townPlots;
-				break;
-			}
+		// 	let filteredPlotList = [];
+		// 	switch(plotStatus){
+		// 		case "Owned":
+		// 			filteredPlotList = townPlots.filter(plot => hasOwner(plot));
+		// 		break;
+		// 		case "Closed":
+		// 			filteredPlotList = townPlots.filter(plot => isClosed(plot) && !hasOwner(plot));
+		// 		break;
+		// 		case "Open":
+		// 			filteredPlotList = townPlots.filter(plot => !isClosed(plot) && !hasOwner(plot));
+		// 		break;
+		// 		case "Built":
+		// 			filteredPlotList = townPlots.filter(plot => hasBuild(plot));
+		// 		break;
+		// 		case "Empty":
+		// 			filteredPlotList = townPlots.filter(plot => !hasBuild(plot) && hasOwner(plot));
+		// 		break;
+		// 		default:
+		// 			filteredPlotList = townPlots;
+		// 		break;
+		// 	}
 
-			return (returnLength) ? filteredPlotList.length : filteredPlotList;
-		}
+		// 	return (returnLength) ? filteredPlotList.length : filteredPlotList;
+		// }
 
-		/**
-		 * This function loads all existing core buildings matching the given towns id.
-		 * @param {object} town Town DB Object
-		 * @returns {Promise <object[]>}
-		 */
-		async function grabTownCoreBuildings(town){
-			const coreTypes = ['grandhall', 'bank', 'market', 'tavern', 'clergy'];
+		// // FUNCTION MOVED 
+		// /**
+		//  * This function loads all existing core buildings matching the given towns id.
+		//  * @param {object} town Town DB Object
+		//  * @param {boolean} [returnLength=false] Set to ``true`` to return the ``.length`` for "built" core buildings
+		//  * @returns {Promise <object[]>}
+		//  */
+		// async function grabTownCoreBuildings(town, returnLength=false){
+		// 	const coreTypes = ['grandhall', 'bank', 'market', 'tavern', 'clergy'];
 
-			const townCoreList = [];
-			for (const type of coreTypes){
-				const core = await CoreBuilding.findOne({where: {townid: town.townid, build_type: type}});
-				if (!core) continue;
-				townCoreList.push(core);
-			}
+		// 	const townCoreList = [];
+		// 	for (const type of coreTypes){
+		// 		const core = await CoreBuilding.findOne({where: {townid: town.townid, build_type: type}});
+		// 		if (!core) continue;
+		// 		townCoreList.push(core);
+		// 	}
 
-			return townCoreList;
-		}
+		// 	return (returnLength) ? townCoreList.length : townCoreList;
+		// }
 
-		/**
-		 * This function handles withdraw logic and updating inventories with applicable amounts and items
-		 * @param {object} town Town DB Object, town giving goods
-		 * @param {object} user UserData DB Object, user recieving goods
-		 * @param {object} navMenu Nav menu object holding all user selected values needed
-		 * @returns {Promise <EmbedBuilder>}
-		 */
-		async function handleWithdrawFromTown(town, user, navMenu){
-			switch(navMenu.typePicked){
-				case "mat":
-					// Give to User
-					await checkInboundMat(user.userid, navMenu.typeExtras.ref, navMenu.typeExtras.matType, navMenu.moveAmount);
-					// Take from Town
-					await checkOutboundTownMat(town.townid, navMenu.typeExtras.ref, navMenu.typeExtras.matType, navMenu.moveAmount);
-				break;
-				case "coin":
-					// Give to User
-					await updateUserCoins(navMenu.moveAmount, user);
-					// Take from Town
-					await spendUserCoins(navMenu.moveAmount, town);
-				break;
-			}
+		// // FUNCTION MOVED 
+		// /**
+		//  * This function handles withdraw logic and updating inventories with applicable amounts and items
+		//  * @param {object} town Town DB Object, town giving goods
+		//  * @param {object} user UserData DB Object, user recieving goods
+		//  * @param {object} navMenu Nav menu object holding all user selected values needed
+		//  * @returns {Promise <EmbedBuilder>}
+		//  */
+		// async function handleWithdrawFromTown(town, user, navMenu){
+		// 	switch(navMenu.typePicked){
+		// 		case "mat":
+		// 			// Give to User
+		// 			await checkInboundMat(user.userid, navMenu.typeExtras.ref, navMenu.typeExtras.matType, navMenu.moveAmount);
+		// 			// Take from Town
+		// 			await checkOutboundTownMat(town.townid, navMenu.typeExtras.ref, navMenu.typeExtras.matType, navMenu.moveAmount);
+		// 		break;
+		// 		case "coin":
+		// 			// Give to User
+		// 			await updateUserCoins(navMenu.moveAmount, user);
+		// 			// Take from Town
+		// 			await spendUserCoins(navMenu.moveAmount, town);
+		// 		break;
+		// 	}
 
-			const finalEmbed = new EmbedBuilder()
-			.setTitle('== Withdraw Successful ==');
+		// 	const finalEmbed = new EmbedBuilder()
+		// 	.setTitle('== Withdraw Successful ==');
 
-			return finalEmbed;
-		}
+		// 	return finalEmbed;
+		// }
 		
-		/**
-		 * This function handles deposit logic and updating inventories with applicable amounts and items
-		 * @param {object} town Town DB Object, town recieving goods
-		 * @param {object} user UserData DB Object, user giving goods
-		 * @param {object} navMenu Nav menu object holding all user selected values needed
-		 * @returns {Promise <EmbedBuilder>}
-		 */
-		async function handleDepositIntoTown(town, user, navMenu){
-			switch(navMenu.typePicked){
-				case "mat":
-					// Give to Town
-					await checkInboundTownMat(town.townid, navMenu.typeExtras.ref, navMenu.typeExtras.matType, navMenu.moveAmount);
-					// Take from user
-					await checkOutboundMat(user.userid, navMenu.typeExtras.ref, navMenu.typeExtras.matType, navMenu.moveAmount);
-				break;
-				case "coin":
-					// Give to town
-					await updateUserCoins(navMenu.moveAmount, town);
-					// take from user
-					await spendUserCoins(navMenu.moveAmount, user);
-				break;
-			}
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles deposit logic and updating inventories with applicable amounts and items
+		//  * @param {object} town Town DB Object, town recieving goods
+		//  * @param {object} user UserData DB Object, user giving goods
+		//  * @param {object} navMenu Nav menu object holding all user selected values needed
+		//  * @returns {Promise <EmbedBuilder>}
+		//  */
+		// async function handleDepositIntoTown(town, user, navMenu){
+		// 	switch(navMenu.typePicked){
+		// 		case "mat":
+		// 			// Give to Town
+		// 			await checkInboundTownMat(town.townid, navMenu.typeExtras.ref, navMenu.typeExtras.matType, navMenu.moveAmount);
+		// 			// Take from user
+		// 			await checkOutboundMat(user.userid, navMenu.typeExtras.ref, navMenu.typeExtras.matType, navMenu.moveAmount);
+		// 		break;
+		// 		case "coin":
+		// 			// Give to town
+		// 			await updateUserCoins(navMenu.moveAmount, town);
+		// 			// take from user
+		// 			await spendUserCoins(navMenu.moveAmount, user);
+		// 		break;
+		// 	}
 
-			const finalEmbed = new EmbedBuilder()
-			.setTitle('== Deposit Successful ==');
+		// 	const finalEmbed = new EmbedBuilder()
+		// 	.setTitle('== Deposit Successful ==');
 
-			return finalEmbed;
-		}
+		// 	return finalEmbed;
+		// }
 
-		/**
-		 * This function handles transfering ownership of the given town, to the given user.
-		 * @param {object} town Town DB Object
-		 * @param {object} user UserData DB Object
-		 * @returns {Promise <{embeds: EmbedBuilder, status: string}>}
-		 */
-		async function updateTownMayor(town, user){
-			await town.update({mayorid: user.userid}).then(async t => await t.save()).then(async t => {return await t.reload()});
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles transfering ownership of the given town, to the given user.
+		//  * @param {object} town Town DB Object
+		//  * @param {object} user UserData DB Object
+		//  * @returns {Promise <{embeds: EmbedBuilder, status: string}>}
+		//  */
+		// async function updateTownMayor(town, user){
+		// 	await town.update({mayorid: user.userid}).then(async t => await t.save()).then(async t => {return await t.reload()});
 
-			const returnEmbed = new EmbedBuilder()
-			.setTitle('== Transfer Complete ==')
-			.setDescription(`${makeCapital(user.username)} has been made the new mayor!`);
+		// 	const returnEmbed = new EmbedBuilder()
+		// 	.setTitle('== Transfer Complete ==')
+		// 	.setDescription(`${makeCapital(user.username)} has been made the new mayor!`);
 
-			return {embeds: [returnEmbed], status: 'Complete'};
-		}
+		// 	return {embeds: [returnEmbed], status: 'Complete'};
+		// }
 
-		/**
-		 * This function handles appointing/demoting a selected user from a given town.
-		 * @param {object} town Town DB Object
-		 * @param {object} user UserData DB Object
-		 * @param {string} changeType One of: ``appoint`` | ``demote``
-		 * @returns {Promise <{embeds: EmbedBuilder, status: string}>}
-		 */
-		async function updateTownCanEditList(town, user, changeType){
-			const curEditList = town.can_edit.split('-');
+		// // FUNCTION MOVED
+		// /**
+		//  * This function handles appointing/demoting a selected user from a given town.
+		//  * @param {object} town Town DB Object
+		//  * @param {object} user UserData DB Object
+		//  * @param {string} changeType One of: ``appoint`` | ``demote``
+		//  * @returns {Promise <{embeds: EmbedBuilder, status: string}>}
+		//  */
+		// async function updateTownCanEditList(town, user, changeType){
+		// 	const curEditList = town.can_edit.split('-');
 
-			const returnEmbed = new EmbedBuilder();
-			let newEditList;
-			switch(changeType){
-				case "appoint":
-					newEditList = curEditList.push(user.userid);
-					returnEmbed
-					.setTitle('== User Appointed ==')
-					.setDescription(`${makeCapital(user.username)} has been appointed to your town!`);
-				break;
-				case "demote":
-					newEditList = curEditList.filter(id => id !== user.userid);
-					returnEmbed
-					.setTitle('== User Demoted ==')
-					.setDescription(`${makeCapital(user.username)} has been demoted from your town!`);
-				break;
-			}
+		// 	const returnEmbed = new EmbedBuilder();
+		// 	let newEditList;
+		// 	switch(changeType){
+		// 		case "appoint":
+		// 			newEditList = curEditList.push(user.userid);
+		// 			returnEmbed
+		// 			.setTitle('== User Appointed ==')
+		// 			.setDescription(`${makeCapital(user.username)} has been appointed to your town!`);
+		// 		break;
+		// 		case "demote":
+		// 			newEditList = curEditList.filter(id => id !== user.userid);
+		// 			returnEmbed
+		// 			.setTitle('== User Demoted ==')
+		// 			.setDescription(`${makeCapital(user.username)} has been demoted from your town!`);
+		// 		break;
+		// 	}
 
-			await town.update({can_edit: newEditList.toString()}).then(async t => await t.save()).then(async t => {return await t.reload()});
+		// 	await town.update({can_edit: newEditList.toString()}).then(async t => await t.save()).then(async t => {return await t.reload()});
 
-			return {embeds: [returnEmbed], status: 'Complete'};
-		}
+		// 	return {embeds: [returnEmbed], status: 'Complete'};
+		// }
 
-		/**
-		 * This function loads the display object for all local town material bonuses available.
-		 * @param {object[]} towns List of all local towns
-		 * @returns {{title: string, color: string, fields: {name: string, value: string}[]}}
-		 */
-		function generateLocalTownBonuses(towns){
-			const finalObj = {
-				title: `== Local Town Bonuses ==`,
-				color: "",
-				fields: []
-			};
+		// // FUNCTION MOVED 
+		// /**
+		//  * This function loads the display object for all local town material bonuses available.
+		//  * @param {object[]} towns List of all local towns
+		//  * @returns {{title: string, color: string, fields: {name: string, value: string}[]}}
+		//  */
+		// function generateLocalTownBonuses(towns){
+		// 	const finalObj = {
+		// 		title: `== Local Town Bonuses ==`,
+		// 		color: "",
+		// 		fields: []
+		// 	};
 
-			const allignMatBonus = {
-				Normal: [10, 5, 3, 3],
-				Evil: [5, 3, 1, 1],
-				Phase: [10, 8, 5, 5],
-			};
+		// 	const allignMatBonus = {
+		// 		Normal: [10, 5, 3, 3],
+		// 		Evil: [5, 3, 1, 1],
+		// 		Phase: [10, 8, 5, 5],
+		// 	};
 
-			const finalFields = [];
-			for (const town of towns){
-				const matTypes = town.mat_bonus.split(',');
-				const bonusList = allignMatBonus[`${town.local_biome.split('-')[1]}`];
+		// 	const finalFields = [];
+		// 	for (const town of towns){
+		// 		const matTypes = town.mat_bonus.split(',');
+		// 		const bonusList = allignMatBonus[`${town.local_biome.split('-')[1]}`];
 
-				let bonusIdx = 0, finalValue = "";
-				for (const type of matTypes){
-					finalValue += `+${bonusList[bonusIdx]} ${makeCapital(type)}\n`;
-					bonusIdx++;
-				}
+		// 		let bonusIdx = 0, finalValue = "";
+		// 		for (const type of matTypes){
+		// 			finalValue += `+${bonusList[bonusIdx]} ${makeCapital(type)}\n`;
+		// 			bonusIdx++;
+		// 		}
 
-				finalFields.push({name: `The Town of ${makeCapital(town.name)} gives: `, value: finalValue});
-			}
+		// 		finalFields.push({name: `The Town of ${makeCapital(town.name)} gives: `, value: finalValue});
+		// 	}
 
-			finalObj.fields = finalFields;
-			finalObj.color = 'DarkGold';
+		// 	finalObj.fields = finalFields;
+		// 	finalObj.color = 'DarkGold';
 
-			return finalObj;
-		}
+		// 	return finalObj;
+		// }
 
-		/**
-		 * This function generates the embed values for the ``town`` requested.
-		 * @param {object} town Town DB Object
-		 * @param {object} mayor UserData DB Object
-		 * @returns {Promise <{title: string, color: string, fields: {name: string, value: string}[]}>}
-		 */
-		async function generateTownDisplayEmbed(town, mayor){
-			const finalObj = {
-				title: `== Town of ${makeCapital(town.name)} ==`,
-				color: "",
-				fields: []
-			};
+		// // FUNCTION MOVED 
+		// /**
+		//  * This function generates the embed values for the ``town`` requested.
+		//  * @param {object} town Town DB Object
+		//  * @param {object} mayor UserData DB Object
+		//  * @returns {Promise <{title: string, color: string, fields: {name: string, value: string}[]}>}
+		//  */
+		// async function generateTownDisplayEmbed(town, mayor){
+		// 	const finalObj = {
+		// 		title: `== Town of ${makeCapital(town.name)} ==`,
+		// 		color: "",
+		// 		fields: []
+		// 	};
 
-			// finalObj.description = `Current Mayor: ${mayor.username}\nTown Level: ${town.level}\nTreasury Contains: ${town.coins}c\n========\nHuman Population: ${town.population}\nNPC Population: ${town.npc_population}\n========\nMax Buildings: ${town.buildlimit}\nOpen Plots: ${town.openplots}\nClosed Plots: ${town.closedplots}\nOwned Plots: ${town.ownedplots}\nBuildings: ${town.buildcount}\n========\nMain Biome: ${biomeList[0]}\nAllignment: ${biomeList[1]}\n========\n`;
+		// 	// finalObj.description = `Current Mayor: ${mayor.username}\nTown Level: ${town.level}\nTreasury Contains: ${town.coins}c\n========\nHuman Population: ${town.population}\nNPC Population: ${town.npc_population}\n========\nMax Buildings: ${town.buildlimit}\nOpen Plots: ${town.openplots}\nClosed Plots: ${town.closedplots}\nOwned Plots: ${town.ownedplots}\nBuildings: ${town.buildcount}\n========\nMain Biome: ${biomeList[0]}\nAllignment: ${biomeList[1]}\n========\n`;
 
-			// Basic Info
-            // ==========
-            // Level, Coins, Location, Population
-            const locationSwitch = town.local_biome.split("-");
-            const basicField = {
-                name: '== Basic Info ==',
-                value: `Town Level: **${town.level}**\nTown Coins: **${town.coins}**c\nTown Biome: **${locationSwitch[1]} ${locationSwitch[0]}**\nPlayer Population: **${town.population}**\nNPC Population: **${town.npc_population}**`
-            };
+		// 	// Basic Info
+        //     // ==========
+        //     // Level, Coins, Location, Population
+        //     const locationSwitch = town.local_biome.split("-");
+        //     const basicField = {
+        //         name: '== Basic Info ==',
+        //         value: `Town Level: **${town.level}**\nTown Coins: **${town.coins}**c\nTown Biome: **${locationSwitch[1]} ${locationSwitch[0]}**\nPlayer Population: **${town.population}**\nNPC Population: **${town.npc_population}**`
+        //     };
 
-            // Mayor Info
-            // ==========
-            // Cur-User?
-            // const theMayor = await grabUser(town.mayorid);
-            const mayorField = {
-                name: '== The Mayor ==',
-                value: `**${makeCapital(mayor.username)}**`
-            };
+        //     // Mayor Info
+        //     // ==========
+        //     // Cur-User?
+        //     // const theMayor = await grabUser(town.mayorid);
+        //     const mayorField = {
+        //         name: '== The Mayor ==',
+        //         value: `**${makeCapital(mayor.username)}**`
+        //     };
 
-            // Build Info
-            // ==========
-            // Tot-Plots, Open, Closed, Built
-            const buildFields = {
-                name: '== Plot Info ==',
-                value: `Total Plots: **${town.buildlimit}**\nOpen Plots: **${town.openplots}**\nClosed Plots: **${town.closedplots}**\nOwned Plots: **${town.ownedplots}**\nDeveloped Plots: **${town.buildcount}**`
-            };
+        //     // Build Info
+        //     // ==========
+        //     // Tot-Plots, Open, Closed, Built
+        //     const buildFields = {
+        //         name: '== Plot Info ==',
+        //         value: `Total Plots: **${town.buildlimit}**\nOpen Plots: **${town.openplots}**\nClosed Plots: **${town.closedplots}**\nOwned Plots: **${town.ownedplots}**\nDeveloped Plots: **${town.buildcount}**`
+        //     };
 
-            // Core Info
-            // =========
-            // Grandhall, Bank, Market, Tavern, Clergy
-			/**
-			 * This method loads the appropete display text given the status of a core building.
-			 * @param {string} status Corebuilding build status
-			 * @returns {string}
-			 */
-			const loadCoreTextFromStatus = (status) => {
-				return (status === 'None') ? "Not Built" : `Progress: ${status.split(': ')[0]} @ ${status.split(': ')[1]}`;
-			};
-			// 0 Grandhall, 1 Bank, 2 Market, 3 Tavern, 4 Clergy
-			const coreTextList = [
-				loadCoreTextFromStatus(town.grandhall_status), 
-				loadCoreTextFromStatus(town.bank_status), 
-				loadCoreTextFromStatus(town.market_status), 
-				loadCoreTextFromStatus(town.tavern_status), 
-				loadCoreTextFromStatus(town.clergy_status)
-			];
-            const coreFields = {
-                name: '== Core-Building Info ==',
-                value: `Grandhall Status: **${coreTextList[0]}**\nBank Status: **${coreTextList[1]}**\nMarket Status: **${coreTextList[2]}**\nTavern Status: **${coreTextList[3]}**\nClergy Status: **${coreTextList[4]}**`
-            };
+        //     // Core Info
+        //     // =========
+        //     // Grandhall, Bank, Market, Tavern, Clergy
+		// 	/**
+		// 	 * This method loads the appropete display text given the status of a core building.
+		// 	 * @param {string} status Corebuilding build status
+		// 	 * @returns {string}
+		// 	 */
+		// 	const loadCoreTextFromStatus = (status) => {
+		// 		return (status === 'None') ? "Not Built" : `Progress: ${status.split(': ')[0]} @ ${status.split(': ')[1]}`;
+		// 	};
+		// 	// 0 Grandhall, 1 Bank, 2 Market, 3 Tavern, 4 Clergy
+		// 	const coreTextList = [
+		// 		loadCoreTextFromStatus(town.grandhall_status), 
+		// 		loadCoreTextFromStatus(town.bank_status), 
+		// 		loadCoreTextFromStatus(town.market_status), 
+		// 		loadCoreTextFromStatus(town.tavern_status), 
+		// 		loadCoreTextFromStatus(town.clergy_status)
+		// 	];
+        //     const coreFields = {
+        //         name: '== Core-Building Info ==',
+        //         value: `Grandhall Status: **${coreTextList[0]}**\nBank Status: **${coreTextList[1]}**\nMarket Status: **${coreTextList[2]}**\nTavern Status: **${coreTextList[3]}**\nClergy Status: **${coreTextList[4]}**`
+        //     };
 
-			// Embed Fields
-			finalObj.fields = [mayorField, basicField, buildFields, coreFields];
+		// 	// Embed Fields
+		// 	finalObj.fields = [mayorField, basicField, buildFields, coreFields];
 
-			const biomeColours = {
-				Forest: 'DarkGreen',
-				Mountain: 'LightGrey',
-				Desert: 'DarkGold',
-				Plains: 'Gold',
-				Swamp: 'DarkAqua',
-				Grassland: 'Green'
-			};
+		// 	const biomeColours = {
+		// 		Forest: 'DarkGreen',
+		// 		Mountain: 'LightGrey',
+		// 		Desert: 'DarkGold',
+		// 		Plains: 'Gold',
+		// 		Swamp: 'DarkAqua',
+		// 		Grassland: 'Green'
+		// 	};
 
-			// Embed Colour
-			finalObj.color = biomeColours[`${locationSwitch[0]}`];
+		// 	// Embed Colour
+		// 	finalObj.color = biomeColours[`${locationSwitch[0]}`];
 
-			return finalObj;
-		}
+		// 	return finalObj;
+		// }
 
-		/**
-		 * This function handles updating both the ``town`` and ``user`` objects given.
-		 * 
-		 * ``town.population++``
-		 * 
-		 * ``user.townid`` = ``town.townid``
-		 * @param {object} town Town DB Object
-		 * @param {object} user UserData DB Object
-		 * @returns {Promise <{embeds: EmbedBuilder, status: string}>}
-		 */
-		async function handleJoinTown(town, user){
-			await town.increment('population').then(async t => await t.save()).then(async t => {return await t.reload()});
-			await handleTownPopLeveling(town);
-			await user.update({townid: town.townid}).then(async u => await u.save()).then(async u => {return await u.reload()});
+		// // FUNCTION MOVED 
+		// /**
+		//  * This function handles updating both the ``town`` and ``user`` objects given.
+		//  * 
+		//  * ``town.population++``
+		//  * 
+		//  * ``user.townid`` = ``town.townid``
+		//  * @param {object} town Town DB Object
+		//  * @param {object} user UserData DB Object
+		//  * @returns {Promise <{embeds: EmbedBuilder, status: string}>}
+		//  */
+		// async function handleJoinTown(town, user){
+		// 	await town.increment('population').then(async t => await t.save()).then(async t => {return await t.reload()});
+		// 	await handleTownPopLeveling(town);
+		// 	await user.update({townid: town.townid}).then(async u => await u.save()).then(async u => {return await u.reload()});
 
-			const finalEmbed = new EmbedBuilder()
-			.setTitle('== Town Joined ==')
-			.setDescription(`Congratulations!! You are now a member of ${makeCapital(town.name)}!`);
+		// 	const finalEmbed = new EmbedBuilder()
+		// 	.setTitle('== Town Joined ==')
+		// 	.setDescription(`Congratulations!! You are now a member of ${makeCapital(town.name)}!`);
 
-			return {embeds: finalEmbed, status: 'Complete'};
-		}
+		// 	return {embeds: finalEmbed, status: 'Complete'};
+		// }
 
-		/**
-		 * This function handles checking if the given ``town`` meets the level up requirements.
-		 * @param {object} town Town DB Object
-		 * @returns {Promise <void>}
-		 */
-		async function handleTownPopLeveling(town){
-			const townLevelCheckStart = new Date().getTime();
-			let timerEndsText = 'Town Level Checking';
+		// // FUNCTION MOVED 
+		// /**
+		//  * This function handles checking if the given ``town`` meets the level up requirements.
+		//  * @param {object} town Town DB Object
+		//  * @returns {Promise <void>}
+		//  */
+		// async function handleTownPopLeveling(town){
+		// 	const townLevelCheckStart = new Date().getTime();
+		// 	let timerEndsText = 'Town Level Checking';
 
-			const staticTownPopLevelReq = (town) => {
-				const totPop = town.population + town.npc_population;
-				const levelUPReq = town.level * 5;
-				return totPop >= levelUPReq;
-			};
+		// 	const staticTownPopLevelReq = (town) => {
+		// 		const totPop = town.population + town.npc_population;
+		// 		const levelUPReq = town.level * 5;
+		// 		return totPop >= levelUPReq;
+		// 	};
 
-			if (staticTownPopLevelReq(town)) {
-				await town.increment('level').then(async t => await t.save()).then(async t => {return await t.reload()});
-				// CHANGE THIS VALUE BASED ON LEVEL/OTHER FACTORS!!
-				const plotsNeeded = 5;
-				await handleNewTownPlotCreation(town, plotsNeeded);
-				timerEndsText += ' (LEVEL UP)';
-			}
+		// 	if (staticTownPopLevelReq(town)) {
+		// 		await town.increment('level').then(async t => await t.save()).then(async t => {return await t.reload()});
+		// 		// CHANGE THIS VALUE BASED ON LEVEL/OTHER FACTORS!!
+		// 		const plotsNeeded = 5;
+		// 		await handleNewTownPlotCreation(town, plotsNeeded);
+		// 		timerEndsText += ' (LEVEL UP)';
+		// 	}
 
-			endTimer(townLevelCheckStart, timerEndsText);
-			return;
-		}
+		// 	endTimer(townLevelCheckStart, timerEndsText);
+		// 	return;
+		// }
 
-		/**
-		 * This function generates ``plotAmount`` new plots linked by ``town.townid``.
-		 * @param {object} town Town DB Object
-		 * @param {number} plotAmount Number of new plots to create
-		 * @returns {Promise <void>}
-		 */
-		async function handleNewTownPlotCreation(town, plotAmount){
-			const makeNewPlotsStart = new Date().getTime();
-			for (let i = 0; i < plotAmount; i++){
-				await TownPlots.create({
-					townid: town.townid
-				}).then(async tp => await tp.save()).then(async tp => {return await tp.reload()});
-			}
+		// // FUNCTION MOVED 
+		// /**
+		//  * This function generates ``plotAmount`` new plots linked by ``town.townid``.
+		//  * @param {object} town Town DB Object
+		//  * @param {number} plotAmount Number of new plots to create
+		//  * @returns {Promise <void>}
+		//  */
+		// async function handleNewTownPlotCreation(town, plotAmount){
+		// 	const makeNewPlotsStart = new Date().getTime();
+		// 	for (let i = 0; i < plotAmount; i++){
+		// 		await TownPlots.create({
+		// 			townid: town.townid
+		// 		}).then(async tp => await tp.save()).then(async tp => {return await tp.reload()});
+		// 	}
 
-			await handleUpdateTownPlotTypeAmounts(town);
+		// 	await handleUpdateTownPlotTypeAmounts(town);
 
-			endTimer(makeNewPlotsStart, 'Create New Town Plots');
-			console.log(`Current Total Plots for ${makeCapital(town.name)}: %d`, await TownPlots.findAll({where: {townid: town.townid}}).length);
-		}
+		// 	endTimer(makeNewPlotsStart, 'Create New Town Plots');
+		// 	console.log(`Current Total Plots for ${makeCapital(town.name)}: %d`, await TownPlots.findAll({where: {townid: town.townid}}).length);
+		// }
 
 
 		// // Establish Town
